Split again so hooks work correctly
